# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "PageEquides": {
        "properties": {
          "donnees": {
            "description": "Résumés par défaut, fiches complètes si `vue=complet`.",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Resume"
                },
                {
                  "$ref": "#/components/schemas/Equide"
                }
              ]
            },
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": [
          "donnees",
          "pagination"
        ],
        "type": "object"
      },
...
      "Probleme": {
        "description": "Corps d'erreur conforme à la RFC 9457 (« Problem Details for HTTP APIs »), servi en `application/problem+json`. Les champs `parametre` et `indice` sont des extensions, que la RFC autorise.",
        "properties": {
          "detail": {
            "description": "Explication propre à cette occurrence.",
            "example": "paramètre inconnu `rase`",
            "type": "string"
          },
          "documentation": {
            "example": "/",
            "type": "string"
          },
          "indice": {
            "description": "Extension : piste de correction — valeurs admises, référentiel à consulter, ou paramètre le plus proche en cas de faute de frappe.",
            "example": "vouliez-vous dire `race` ?",
            "type": "string"
          },
          "parametre": {
            "description": "Extension : paramètre fautif, quand l'erreur en désigne un.",
            "example": "rase",
            "type": "string"
          },
          "status": {
            "example": 400,
            "type": "integer"
          },
          "title": {
            "description": "Résumé lisible, stable pour un `type` donné.",
            "example": "Paramètre invalide",
            "type": "string"
          },
          "type": {
            "description": "URI du type de problème, relative — le client la résout contre l'URI de la requête. Elle pointe vers la section correspondante de la documentation.",
            "example": "/#erreur-parametre-invalide",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Fiches d'origine",
      "url": "https://infochevaux.ifce.fr"
    },
...
    "license": {
      "identifier": "MIT",
      "name": "MIT (code)"
    },
    "summary": "Interrogation des données publiques des équidés enregistrés en France depuis 1976.",
    "title": "API Équidés",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/openapi.json": {
      "get": {
        "description": "La page de documentation est rendue depuis ce document : l'une ne peut pas diverger de l'autre.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Document OpenAPI 3.1."
          },
          "304": {
            "description": "Contenu inchangé depuis l'`ETag` fourni en `If-None-Match`. Aucun corps, et rien n'est recalculé côté service."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "Quota de débit dépassé pour l'adresse appelante. `Retry-After` indique le délai."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "Capacité de traitement simultané atteinte, délai maximal dépassé, ou données non chargées. `Retry-After` indique le délai."
          }
        },
        "summary": "Description OpenAPI de ce service",
        "tags": [
          "Service"
        ]
      }
    },
...
    "/v1/equides": {
      "get": {
        "description": "Les filtres d'une même dimension se combinent en OU, les dimensions différentes en ET.\n\nLes lignes sont renvoyées en **résumé** par défaut ; `vue=complet` sert les fiches entières.\n\nAucune réponse ne porte de lien déductible d'un identifiant : les gabarits d'URL sont publiés par `GET /`.",
        "parameters": [
          {
            "description": "Race, répétable pour un OU (`?race=A&race=B`) ou séparée par des virgules. Insensible à la casse et aux accents. Le jeu compte 11 747 races distinctes ; valeurs admises sur /v1/referentiels/races.",
            "example": "Pur Sang",
            "in": "query",
            "name": "race",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "description": "Robe, mêmes règles que `race`. Valeurs admises : /v1/referentiels/robes.",
            "example": "Bai",
            "in": "query",
            "name": "robe",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "description": "Sexe tel qu'écrit par la source : `Femelle`, `Male`, `Hongre`, `Indeter`.",
            "example": "Femelle",
            "in": "query",
            "name": "sexe",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "description": "Restreint aux équidés portant des performances dans cette discipline. Valeurs admises : /v1/referentiels/disciplines.",
            "example": "TROT COURSE",
            "in": "query",
            "name": "discipline",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "description": "Restreint aux équidés portant un indice de ce code (`BTR`, `ISO`, `ITR`…). Valeurs admises : /v1/referentiels/codes_indice.",
            "example": "ISO",
            "in": "query",
            "name": "indice",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "description": "Millésime de naissance, répétable.",
            "example": 2015,
            "in": "query",
            "name": "annee_naissance",
            "required": false,
            "schema": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            }
          },
          {
            "description": "Recherche sur le nom. Chaque mot de la requête est traité comme un préfixe et tous doivent être présents : `qabalah mer` retrouve « QABALAH MERCURY ».",
            "example": "qabalah",
            "in": "query",
            "name": "nom",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Millésime de naissance minimal, inclus.",
            "example": 2010,
            "in": "query",
            "name": "annee_min",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Millésime de naissance maximal, inclus.",
            "example": 2020,
            "in": "query",
            "name": "annee_max",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Restreint aux équidés portant (ou non) des indices de performance. Accepte `true`, `false`, `1`, `0`, `oui`, `non`. Attention : `false` ne signifie pas que l'équidé n'a jamais concouru, seulement qu'aucun indice n'est publié — 75,1 % du jeu est dans ce cas.",
            "example": true,
            "in": "query",
            "name": "avec_performances",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Ordre : `naturel` (ordre du fichier source, défaut), `nom`, `annee`. Préfixer par `-` inverse le sens. Les millésimes absents sont toujours placés en fin de liste.",
            "example": "-annee",
            "in": "query",
            "name": "tri",
            "required": false,
            "schema": {
              "default": "naturel",
              "enum": [
                "naturel",
                "nom",
                "annee",
                "-naturel",
                "-nom",
                "-annee"
              ],
              "type": "string"
            }
          },
          {
            "description": "Niveau de détail des lignes renvoyées : `resume` (défaut, ~180 octets par ligne) ou `complet` (fiche entière, ~1,1 Kio par ligne).",
            "example": "resume",
            "in": "query",
            "name": "vue",
            "required": false,
            "schema": {
              "default": "resume",
              "enum": [
                "resume",
                "complet"
              ],
              "type": "string"
            }
          },
          {
            "description": "Lignes par page (1 à 100, défaut 20). Une valeur hors bornes est refusée par un `400`.",
            "example": 20,
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Décalage, plafonné à 10000 ; au-delà, un `400`. La pagination profonde dégrade le service : pour parcourir le jeu entier, téléchargez le fichier publié.",
            "example": 0,
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 10000,
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageEquides"
                }
              }
            },
            "description": "Page de résultats."
          },
          "304": {
            "description": "Contenu inchangé depuis l'`ETag` fourni en `If-None-Match`. Aucun corps, et rien n'est recalculé côté service."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "Paramètre inconnu, valeur hors bornes, ou modalité que le référentiel n'admet pas."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "Quota de débit dépassé pour l'adresse appelante. `Retry-After` indique le délai."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "Capacité de traitement simultané atteinte, délai maximal dépassé, ou données non chargées. `Retry-After` indique le délai."
          }
        },
        "summary": "Rechercher des équidés",
        "tags": [
          "Équidés"
        ]
      }
    },
    "/v1/equides/{id}": {
      "get": {
        "parameters": [
          {
            "description": "Identifiant renvoyé par la recherche : celui de l'IFCE, 22 caractères en base64 URL.",
            "example": "Z4ogLhlkS2CeUdq0bZ0YFw",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equide"
                }
              }
            },
            "description": "L'équidé demandé."
          },
          "304": {
            "description": "Contenu inchangé depuis l'`ETag` fourni en `If-None-Match`. Aucun corps, et rien n'est recalculé côté service."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "Identifiant inconnu ou mal formé — les deux cas donnent la même réponse."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "Quota de débit dépassé pour l'adresse appelante. `Retry-After` indique le délai."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "Capacité de traitement simultané atteinte, délai maximal dépassé, ou données non chargées. `Retry-After` indique le délai."
          }
        },
        "summary": "Consulter un équidé",
        "tags": [
          "Équidés"
        ]
      }
    },
...
    "/v1/meta": {
      "get": {
        "description": "La réutilisation impose de mentionner la source et la date de mise à jour : `source`, `ingere_le` et `empreinte_source` y pourvoient.\n\nLes `remarques` énoncent les défauts connus de la source et le régime de réutilisation applicable.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meta"
                }
              }
            },
            "description": "Métadonnées du jeu de données chargé."
          },
          "304": {
            "description": "Contenu inchangé depuis l'`ETag` fourni en `If-None-Match`. Aucun corps, et rien n'est recalculé côté service."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "Quota de débit dépassé pour l'adresse appelante. `Retry-After` indique le délai."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "Capacité de traitement simultané atteinte, délai maximal dépassé, ou données non chargées. `Retry-After` indique le délai."
          }
        },
        "summary": "Provenance, licence et qualité de la donnée servie",
        "tags": [
          "Service"
        ]
      }
    },
    "/v1/referentiels/{dimension}": {
      "get": {
        "description": "Sans cette liste, un intégrateur ne peut pas deviner que la source écrit `Trotteur Francais` sans cédille ni accent, ni que le sexe se dit `Femelle` et non `F`.\n\nLes modalités sont rendues de la plus fréquente à la plus rare, et la liste est tronquée à 2000 entrées — `nombre_de_modalites` donne alors le total réel et `note` signale la troncature. `annees_naissance` fait exception : rendu en entier, par ordre chronologique.",
        "parameters": [
          {
            "description": "Référentiel demandé. `races_lien` porte les 13 races propres déclarées sur les liens de filiation, là où `races` reflète un champ pollué par la source.",
            "example": "races",
            "in": "path",
            "name": "dimension",
            "required": true,
            "schema": {
              "enum": [
                "races",
                "robes",
                "sexes",
                "disciplines",
                "codes_indice",
                "races_lien",
                "annees_naissance"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Referentiel"
                }
              }
            },
            "description": "Modalités et effectifs, du plus fréquent au plus rare."
          },
          "304": {
            "description": "Contenu inchangé depuis l'`ETag` fourni en `If-None-Match`. Aucun corps, et rien n'est recalculé côté service."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "Référentiel inconnu."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "Quota de débit dépassé pour l'adresse appelante. `Retry-After` indique le délai."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "Capacité de traitement simultané atteinte, délai maximal dépassé, ou données non chargées. `Retry-After` indique le délai."
          }
        },
        "summary": "Lister les valeurs admises par un filtre",
        "tags": [
          "Référentiels"
        ]
      }
    },
    "/v1/search": {
      "get": {
        "description": "Renvoie des suggestions **classées par pertinence**, pour alimenter un champ de saisie.\n\nÀ distinguer de `/v1/equides?nom=…`, qui filtre et pagine sans classer. Ici le coût est borné par `limite` et non par la popularité du préfixe : la réponse doit arriver avant la frappe suivante.",
        "parameters": [
          {
            "description": "Début de nom, insensible à la casse et aux accents. Obligatoire et non vide.",
            "example": "qab",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "Nombre de suggestions (1 à 100, défaut 10). Une valeur hors bornes est ramenée au plafond, non refusée.",
            "example": 10,
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Suggestions"
                }
              }
            },
            "description": "Suggestions, de la plus pertinente à la moins."
          },
          "304": {
            "description": "Contenu inchangé depuis l'`ETag` fourni en `If-None-Match`. Aucun corps, et rien n'est recalculé côté service."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "`q` manquant ou vide, ou paramètre inconnu."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "Quota de débit dépassé pour l'adresse appelante. `Retry-After` indique le délai."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Probleme"
                }
              }
            },
            "description": "Capacité de traitement simultané atteinte, délai maximal dépassé, ou données non chargées. `Retry-After` indique le délai."
          }
        },
        "summary": "Autocomplétion sur le nom",
        "tags": [
          "Équidés"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Ce serveur",
      "url": "/"
    }
  ],
...
}